Refactor generateAddTodoBtn
